requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-huggingface
chromadb==0.4.24
rank-bm25
sentence-transformers
pypdf
unstructured[docx,md]
pdf2image
pytesseract
Pillow
pydantic
streamlit